Add dialog for color modes
