+++ src/dialogs/more-info/components/ha-more-info-light-hs.ts
@@ -1,4 +1,3 @@
-import { mdiPalette } from "@mdi/js";
 import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import "../../../components/ha-icon-button";
@@ -6,45 +5,34 @@ import { LightEntity } from "../../../data/light";
 import { HomeAssistant } from "../../../types";
 import "./ha-more-info-bar-slider";
 
-@customElement("ha-more-info-light-color")
-export class HaMoreInfoLightColor extends LitElement {
+@customElement("ha-more-info-light-hs")
+export class HaMoreInfoLightHS extends LitElement {
   @property({ attribute: false }) public hass!: HomeAssistant;
 
   @property({ attribute: false }) public stateObj!: LightEntity;
 
-  @state() value?: [number, number, number];
-
-  @state() private _hueSegments = 24;
-
-  @state() private _saturationSegments = 8;
+  @state() value?: { h: number; s: number };
 
   protected updated(changedProp: Map<string | number | symbol, unknown>): void {
     if (changedProp.has("stateObj")) {
-      this.value = this.stateObj?.attributes?.rgb_color;
-    }
-  }
-
-  private _segmentClick() {
-    if (this._hueSegments === 24 && this._saturationSegments === 8) {
-      this._hueSegments = 0;
-      this._saturationSegments = 0;
-    } else {
-      this._hueSegments = 24;
-      this._saturationSegments = 8;
+      this.value = {
+        h: this.stateObj!.attributes.hs_color![0],
+        s: this.stateObj!.attributes.hs_color![1] / 100,
+      };
     }
   }
 
   private _colorPicked(
     ev: CustomEvent<{
       hs: { h: number; s: number };
-      rgb: { r: number; g: number; b: number };
+      rgb: [number, number, number];
     }>
   ) {
-    const value = [ev.detail.rgb.r, ev.detail.rgb.g, ev.detail.rgb.b];
+    const value = [ev.detail.hs.h, ev.detail.hs.s * 100];
 
     this.hass.callService("light", "turn_on", {
       entity_id: this.stateObj!.entity_id,
-      rgb_color: value,
+      hs_color: value,
     });
   }
 
@@ -54,17 +42,10 @@ export class HaMoreInfoLightColor extends LitElement {
         <ha-color-picker
           class="color"
           @colorselected=${this._colorPicked}
-          .desiredRgbColor=${this.value}
-          throttle="500"
-          .hueSegments=${this._hueSegments}
-          .saturationSegments=${this._saturationSegments}
+          .desiredHsColor=${this.value}
+          .throttle=${500}
         >
         </ha-color-picker>
-        <ha-icon-button
-          .path=${mdiPalette}
-          @click=${this._segmentClick}
-          class="button"
-        ></ha-icon-button>
         <p>Color</p>
       </div>
     `;
@@ -87,13 +68,8 @@ export class HaMoreInfoLightColor extends LitElement {
         --ha-color-picker-wheel-shadow: none;
         --ha-color-picker-marker-borderwidth: 2;
         --ha-color-picker-marker-bordercolor: white;
-      }
-
-      .button {
-        position: absolute;
-        top: 5%;
-        left: 0;
-        color: var(--secondary-text-color);
+        height: 320px;
+        padding-top: 52px;
       }
 
       p {
@@ -109,6 +85,6 @@ export class HaMoreInfoLightColor extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    "ha-more-info-light-color": HaMoreInfoLightColor;
+    "ha-more-info-light-hs": HaMoreInfoLightHS;
   }
 }
